MOB-1614 - Communities UI (#270)
* Added separation on communities section

* Implemented UI for community list item

* Implemented formatting of number of badge holders

* Handle join community action

* Implemented empty state and header UI for community chat

* Updated Ui for unread messages indicator

* Check for UD Blue membership before join the community

* Refresh UDBlue status on app active

* Pull up service refactoring

* Fixed typo

* Show public domain profile on tap of community user

* Switch to communities segment when necessary

* Fixed threads issue

* Fixed unread messages indicator for communities

* Implemented UI for communities empty state

* Finished UI of 'Enable messaging profile' state

* Show toast when messaging profile created

* Cleared TODOs

* Fixed compile error

* Close chat when user left the community.
Handle tap on community cell

* Reset logs topics

* Include communities into a search results

* Disable images functionality for communities

* Updated search for communities rules
