+++ unstoppable-ios-app/domains-manager-ios/Extensions/Double.swift
@@ -10,3 +10,13 @@ import Foundation
 extension Double {
     var ethValue: Double { self / 1_000_000_000_000_000_000 }
 }
+
+extension Double {
+    func reduceScale(to places: Int) -> Double {
+        let multiplier = pow(10, Double(places))
+        let newDecimal = multiplier * self // move the decimal right
+        let truncated = Double(Int(newDecimal)) // drop the fraction
+        let originalDecimal = truncated / multiplier // move the decimal back
+        return originalDecimal
+    }
+}
