+++ unstoppable-ios-app/domains-manager-ios/Modules/Messaging/Entities/Chats/MessagingCommunitiesChatDetails.swift
@@ -15,10 +15,12 @@ struct MessagingCommunitiesChatDetails: Hashable, Codable {
     let pendingMembers: [MessagingChatUserDisplayInfo]
     let adminWallets: [String]
     
+    var allMembers: [MessagingChatUserDisplayInfo] { members + pendingMembers }
+
     var displayName: String {
         switch type {
         case .badge(let badge):
-            return (isJoined ? "✅" : "🟥") + badge.badge.name
+            return badge.badge.name
         }
     }
